restructure http clients
