+++ ComicEngine.Graphql/ApplicationConfiguration.cs
@@ -1,6 +1,3 @@
-using ComicEngine.Graphql.ComicEngineApi;
-using ComicEngine.Graphql.HttpClients;
-
 namespace ComicEngine.Graphql {
     public class ApplicationConfiguration {
         // configuration for API Http Client.
