+++ ComicEngine.Graphql/ComicHttpClientConfig.cs
@@ -1,4 +1,4 @@
-namespace ComicEngine.Graphql.HttpClients {
+namespace ComicEngine.Graphql {
     /// <summary>
     /// Configuration Object for the current in use ComicHttpClient, This will contain any special 
     /// requirements, as well as the url string to send requests.
